backend/requirements.txt
fastapi==0.115.14
psycopg2-binary==2.9.10
asyncpg==0.30.0
sqlalchemy==2.0.38
uvicorn==0.32.1
python-dotenv==1.0.0
websockets==13.1